requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
PyPDF2==3.0.1
python-dotenv==1.0.1
aiofiles==23.2.1
Pillow==10.2.0
pdf2image==1.16.3
pytesseract==0.3.10 